requirements.txt
asgiref==3.2.10
Django==3.1.13
djangorestframework==3.11.2
numpy==1.19.0
opencv-python==4.2.0.34
pytz==2020.1
sqlparse==0.3.1
whitenoise